Update common conf files for play 2.4.0
