+++ common/app/implicits/WSRequests.scala
@@ -2,13 +2,13 @@ package implicits
 
 import common.HttpStatusException
 import conf.Configuration
-import play.api.libs.ws.{WSAuthScheme, WSRequestHolder, WSResponse}
+import play.api.libs.ws.{WSAuthScheme, WSRequest, WSResponse}
 
 import scala.concurrent.{ExecutionContext, Future}
 
 trait WSRequests {
 
-  implicit class RichWSRequestHolder(wsRequest: WSRequestHolder) {
+  implicit class RichWSRequest(wsRequest: WSRequest) {
 
     def getOKResponse()(implicit ec: ExecutionContext): Future[WSResponse] = {
       wsRequest.get() flatMap { response =>
@@ -19,7 +19,7 @@ trait WSRequests {
       }
     }
 
-    def withPreviewAuth: WSRequestHolder = Configuration.contentApi.previewAuth
+    def withPreviewAuth: WSRequest = Configuration.contentApi.previewAuth
       .foldLeft(wsRequest){ case (r, auth) => r.withAuth(auth.user, auth.password, WSAuthScheme.BASIC)}
   }
 }
